perf: speed up rotation for special angles
I found a large amount of time is spent in DkImage::rotateImage() when nomacs
is opening image in a directory of many large (10000 x 10000) images,
with the thumbnails panel on.

For angles 0, 90, 180, and 270, we can use matrix operations instead of
painting to speed up the process. Also, because the downstream
processing does not seem to require the format to be RGBA8888, we can
also skip the costly format conversion.

This commit also adds a benchmark target, which might be helpful for
tuning and comparing performance in the future.